enable multiple services running with different config files
