+++ config-addressbook.yml
@@ -1,0 +1,14 @@
+application:
+  debug: True
+  environment: 'dev'
+  logfile: '/tmp/byoda-addressbook.log'
+  network: 'byoda.net'
+  jaeger_host: 192.168.1.11
+
+svcserver:
+  name: 'serviceserver'
+  root_dir: '/byoda'
+  roles:
+    - service
+  service_id: 1
+  private_key_password: 'byoda'
